refactor: :technologist: change Footer component to Smart and Dumb component concept
